Improved adaptive dt handling
It is now possible to set a different time grid for each regime in the multi-regime pricer. The default values are sufficient for precise simulations, especially for short regimes that require a very small timescale.

The random grid simulation is now optimized to reuse Monte Carlo simulations for regimes with the same timescale, instead of generating data for each smile individually.